API-12088 remove common objects from configuration and adjust structures
